Move Entity creation from AppDatabase to the Entitiys Dao
