+++ app/src/main/java/com/paunoski/manuel/todoapp/MainActivity.java
@@ -11,8 +11,8 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
 
-import com.paunoski.manuel.todoapp.db.AppDatabase;
 import com.paunoski.manuel.todoapp.db.Todo;
+import com.paunoski.manuel.todoapp.db.TodoDao;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -30,7 +30,7 @@ public class MainActivity extends AppCompatActivity {
         list = new ArrayList<>();
         String s = "Todo #";
         for (int i = 0; i < 3; i++) {
-            Todo todo = AppDatabase.createTodo(this);
+            Todo todo = TodoDao.createTodo(this);
             todo.text = s + i;
             list.add(todo);
         }
@@ -48,7 +48,7 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void submit(View view) {
-        Todo todo = AppDatabase.createTodo(this);
+        Todo todo = TodoDao.createTodo(this);
         todo.text = editText.getText().toString();
         list.add(todo);
         adapter.notifyItemInserted(list.size() - 1);
