+++ app/src/main/java/com/paunoski/manuel/todoapp/db/AppDatabase.java
@@ -7,13 +7,10 @@ import android.arch.persistence.room.RoomDatabase;
 import android.content.Context;
 import android.support.annotation.VisibleForTesting;
 
-import java.util.concurrent.atomic.AtomicInteger;
-
 
 @Database(entities = {Todo.class}, version = 1)
 public abstract class AppDatabase extends RoomDatabase {
     private static AppDatabase instance;
-    private static AtomicInteger nextId;
 
     public static synchronized AppDatabase getInstance(Context context) {
         if (instance == null) {
@@ -27,19 +24,6 @@ public abstract class AppDatabase extends RoomDatabase {
     @VisibleForTesting
     public static void switchToInMemory(Context context) {
         instance = Room.inMemoryDatabaseBuilder(context, AppDatabase.class).build();
-    }
-
-    private static int getNextId(Context context) {
-        if (nextId == null) {
-            nextId = new AtomicInteger(getInstance(context).todoDao().getId());
-        }
-        return nextId.incrementAndGet();
-    }
-
-    public static Todo createTodo(Context context) {
-        Todo todo = new Todo();
-        todo.id = getNextId(context);
-        return todo;
     }
 
     public abstract TodoDao todoDao();
