+++ app/src/main/java/com/paunoski/manuel/todoapp/db/TodoDao.java
@@ -3,17 +3,35 @@ package com.paunoski.manuel.todoapp.db;
 import android.arch.persistence.room.Dao;
 import android.arch.persistence.room.Insert;
 import android.arch.persistence.room.Query;
+import android.content.Context;
 
 import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
 
 @Dao
-public interface TodoDao {
+public abstract class TodoDao {
+    private static AtomicInteger nextId;
+
+    private static int getNextId(Context context) {
+        if (nextId == null) {
+            nextId = new AtomicInteger(AppDatabase.getInstance(context).todoDao().getId());
+        }
+        return nextId.incrementAndGet();
+    }
+
+    public static Todo createTodo(Context context) {
+        Todo todo = new Todo();
+        todo.id = getNextId(context);
+        return todo;
+    }
+
     @Query("SELECT * FROM todo")
-    List<Todo> getAll();
+    public abstract List<Todo> getAll();
 
     @Insert
-    void insertAll(Todo... todos);
+    public abstract void insertAll(Todo... todos);
 
     @Query("SELECT MAX(id) FROM todo")
-    int getId();
+    public abstract int getId();
+
 }
